+++ flexiapi/database/factories/PasswordFactory.php
@@ -30,11 +30,10 @@ class PasswordFactory extends Factory
     public function definition()
     {
         $account = Account::factory()->create();
-        $realm = space()?->account_realm ?? $account->domain;
 
         return [
             'account_id' => $account->id,
-            'password'   => hash('md5', $account->username.':'.$realm.':testtest'),
+            'password'   => hash('md5', $account->username.':'.$account->resolvedRealm.':testtest'),
             'algorithm'  => 'MD5',
         ];
     }
@@ -54,10 +53,9 @@ class PasswordFactory extends Factory
     {
         return $this->state(function (array $attributes) {
             $account = Account::find($attributes['account_id']);
-            $realm = space()?->account_realm ?? $account->domain;
 
             return [
-                'password'   => hash('sha256', $account->username.':'.$realm.':testtest'),
+                'password'   => hash('sha256', $account->username.':'.$account->resolvedRealm.':testtest'),
                 'account_id' => $account->id,
                 'algorithm'  => 'SHA-256',
             ];